refactor: Переименовал constants -> data . Также избавился от магических чисел
