+++ NFT-Marketplace/src/components/Creators.jsx
@@ -1,4 +1,4 @@
-import { creatorsCard } from '../constants';
+import { breakPoints, creatorsCard } from '../data';
 import Button from './button';
 import buttonIcon from '../assets/RocketLaunchOutlined.svg';
 import { chopArray, getImage } from '../helpers';
@@ -8,13 +8,13 @@ const Creators = () => {
   const [width] = useWindowSize();
   let cardsArray = [];
 
-  if (width >= 0) {
+  if (width >= breakPoints.start) {
     cardsArray = chopArray(creatorsCard, 5);
   }
-  if (width >= 834) {
+  if (width >= breakPoints.preMiddle) {
     cardsArray = chopArray(creatorsCard, 6);
   }
-  if (width >= 1150) {
+  if (width >= breakPoints.large) {
     cardsArray = creatorsCard;
   }
