+++ NFT-Marketplace/src/components/Trending.jsx
@@ -1,18 +1,18 @@
 import { chopArray, getImage } from '../helpers';
-import { trendingCollections } from '../constants';
+import { breakPoints, trendingCollections } from '../data';
 import { useWindowSize } from '../hooks/useWindowSize';
 
 const Trending = () => {
   const [width] = useWindowSize();
   let newCollection;
 
-  if (width >= 0) {
-    newCollection = chopArray(trendingCollections, 1);
+  if (width >= breakPoints.start) {
+    newCollection = chopArray(trendingCollections, 1)
   }
-  if (width >= 680) {
+  if (width >= breakPoints.small) {
     newCollection = chopArray(trendingCollections, 2);
   }
-  if (width >= 866) {
+  if (width >= breakPoints.middle) {
     newCollection = trendingCollections;
   }
